requirements.txt
Flask
flask-cors
ffmpeg-python